trunks: enable binder.
This includes a trunksd.rc file which starts the service as root,
seccomp policies, and a few other changes to get trunks working on
brillo.

BUG=25360556
TEST=unit tests on target
TEST=manual test on all brilloemulator*-eng boards

+++ trunks/policy_session_test.cc
@@ -109,7 +109,7 @@ TEST_F(PolicySessionTest, GetDigestSuccess) {
       .WillOnce(DoAll(SetArgPointee<2>(policy_digest),
                       Return(TPM_RC_SUCCESS)));
   EXPECT_EQ(TPM_RC_SUCCESS, session.GetDigest(&digest));
-  EXPECT_EQ(SHA256_DIGEST_SIZE, digest.size());
+  EXPECT_EQ(static_cast<size_t>(SHA256_DIGEST_SIZE), digest.size());
 }
 
 TEST_F(PolicySessionTest, GetDigestFailure) {
@@ -167,7 +167,7 @@ TEST_F(PolicySessionTest, PolicyPCRSuccess) {
   EXPECT_EQ(TPM_RC_SUCCESS, session.PolicyPCR(pcr_index, pcr_digest));
   uint8_t pcr_select_index = pcr_index / 8;
   uint8_t pcr_select_byte = 1 << (pcr_index % 8);
-  EXPECT_EQ(pcr_select.count, 1);
+  EXPECT_EQ(pcr_select.count, 1u);
   EXPECT_EQ(pcr_select.pcr_selections[0].hash, TPM_ALG_SHA256);
   EXPECT_EQ(pcr_select.pcr_selections[0].sizeof_select, PCR_SELECT_MIN);
   EXPECT_EQ(pcr_select.pcr_selections[0].pcr_select[pcr_select_index],
